requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.5
python-dotenv==1.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.31
